requirements.txt
cvxpy
ipdb
ipython
jupyter
matplotlib
msprime
numpy
pytest
pyvcf
scikit-allel
scipy
sklearn
tabulate
